Triangle shape option, upload image option (not functional)
I added a triangle shape option, and attempted to allow users to upload
an image from their computer. I tried using a file input and then an
upload button that uses a function that gets the information from the
input and draws the image onto the sketchpad. I don't think the fil
input will work so I am looking for other options