Compacting, Refactoring, and Reorganizing the Quiz
Condensed most of the code for the quiz into a single html page called
"Main.html".  All major activity will be done here, with the use of
<div> instead of frames.  This makes for better readability without
having to switch between dozens of files, as well as the ease of use of
updating information in other sections of the page (unable to do with
frames).

All questions and answers are entered into arrays, and these update the
<span>s that hold the information on the main page.  The user's score
gets updated and the correct answer is displayed after they select their
choice.