Fixed the problem with the actuators
So.. Executive decision time here.  We "shouldn't" use PWM with the
linear actuators.  This means no speed control, I know.  After some
testing, I realized that when we were sending over our voltage signal
with PWM, at the very edges of the range (positions 0-5 and 95-100) the
voltage was very low, not enough to drive the motor in the actuator.
This is why we were hearing the buzzing, the motor was trying to push
out but it couldn't because it didn't have enough voltage to do so.  So,
yes we could increase the voltage supplied at the ends and still use
PWM, but in my opinion, just suppling the full 12 volts all the time
seems like the best option.  We can discuss more today at 2:00.

TL/DR: Not enough voltage supplied to actuators using PWM, should just
push straight 12 Volts all the time.